test(event): consolidate eventually helpers into helpers_test.go
The root event package had four deterministic polling helpers scattered
across ack_policy_test.go, coalesce_test.go, and event_test.go:

  eventuallyEqInt32     (ack_policy_test.go)
  eventuallyTrue        (event_test.go)
  consistentlyEqInt32   (event_test.go)
  waitInputsHandled     (coalesce_test.go)

Each ended up where it was first introduced, but all are general-purpose
polling primitives usable across the package's tests. Move them into a
new helpers_test.go alongside a header that explains why the root
package can't share the testutil.Eventually helper directly.

No behavior changes; the helper bodies are byte-for-byte identical.

go test -race -count=10 passes for the root package.
